+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # rollup changelog
 
+## 1.12.1
+*2019-05-16*
+
+### Bug Fixes
+* Extend file name sanitation to also replace "?" and "*" e.g. when preserving modules with the updated commonjs plugin (#2860)
+* Do not ignore module transformer that return an empty string (#2861)
+
+### Pull Requests
+* [#2860](https://github.com/rollup/rollup/pull/2860): Update to latest plugins and extend file name sanitation (@lukastaegert)
+* [#2861](https://github.com/rollup/rollup/pull/2861): Allow transformers to return an empty string (@lukastaegert)
+
 ## 1.12.0
 *2019-05-15*